// EscalafonFrontend/src/app/interfaces/hoja-vida.ts
export interface HojaVida {
    /**
     * @type { number } - iHojVidId
     */
    iHojVidId?: number;

    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

    /**
     * @type { string } - cHojVidAnotaciones
     */
    cHojVidAnotaciones?: string;

    /**
     * @type { string } - cHojVidNumeroExpediente
     */
    cHojVidNumeroExpediente?: string;

    /**
     * @type { string } - cHojVidNombreSolicitante
     */
    cHojVidNombreSolicitante?: string;

    /**
     * @type { Date } - dHojVidFechaExpediente
     */
    dHojVidFechaExpediente?: Date;

    /**
     * @type { Date } - dHojVidFechaGeneracion
     */
    dHojVidFechaGeneracion?: Date;

    /**
     * @type { string } - cHojVidUuid
     */
    cHojVidUuid?: string;

    /**
     * @type { string } - cHojVidNombreArchivo
     */
    cHojVidNombreArchivo?: string;

}